[SC-216614] Update synced-segment manifests to include cohortId per member
When validating that all users in a batch are coming from the same
cohort, we parse the cohortId from the top-level `cohortIdPath` if a
`cohortIdPath` is not provided in the `memberParser` for synced
segments. Using the top level path applies the same cohortId to all
members, which can result in incorrect data if the members are in fact
from different cohorts. This updates the integration manifests that
define synced-segments to include a `cohortIdPath` and `cohortNamePath`
in the `memberParser` so that cohorts can be validated for all members.