[RO-471] Add PagerDuty integration manifest
This adds the integration manifest for the new PagerDuty - Release
Guardian integration. Right now the templates for the payloads to PD are
stubbed, to be completed in another ticket.

I realized that other integrations may not want to have form variables
to filter their events by, so removed the `triggers` section that was
meant to set those form variables up and used the new `multiselect` and
`environment select` types for form variables for this integration.